Add 'scripts/gc.sh' to watch_controller function in 'scripts/start.py'
The 'start.py' script monitors the seafile-server in a 'watch_controller' function and terminates the container when the server crashed.
However, during a garbage collector cleanup the server needs to be shut down and therefore it is necessary that the 'watch_controller' function only terminates the container if the server is offline and no cleanup is in progress.